Clean up and consolidate migration files
Delete extraneous migrations and making it so that there's exactly one
migration file per DB table. This required changing the order of the
migrations slightly so that every tables is created before any table
that might have foreign keys pointing to it.

This makes things a lot cleaner, at the expense of messing up any
existing deployments that have a schema_migrations table which doesn't
match these migrations. But since we don't *have* any existing
deployments yet, I think we can get away with this ;)