rename 'typeahead' dir to 'Typeahead'
This should hopefully stop all hell from breaking loose when
we upload to Heroku (because browserify tries to require
the module with the capitalized name).

(Note that this doesn't actually change anything on my
filesystem; my local copy of the dir was already capitalized.
Git had just messed it up because 'ignorecase' was 'false'
in my .git/config)