Add Account#onboarding_stage enum
This simplifies the process of checking whether or not a user has
completed the onboarding stage, and, if they haven't completed it,
knowing which page to redirect them to. (Previously the app had to make
2 or 3 extra SQL calls on every page load just to check if the currently
logged-in account needed to be redirected to the survey.

This also means we can get rid of Passenger#has_added_balances and
Passenger#has_added_cards; they're now redundant.