major refactoring of SpendingInfo and its survey
We've decided that 'shared spending' will ALWAYS be true - if you don't
share spending with someone, they should create their own account, not
be part of yours.

The concept of a 'passenger' has been narrows. Really, you should be
able to add *people* to your account (Passenger was already renamed to
Person in a previous commit) - some of these people may be 'card
earners', others might not apply for cards but still want to have their
balances tracked; others may just be travel companions with no cards or
balances that you care about.

Finally, pretty much every page in the survey should also be available
outside of the normal 'survey flow' - e.g. you should be able to add a
person to your account at any time, even if you've already been
onboarded. So the current survey system with all its redirects is too
restrictive and narrow in scope.

As a result of this we're heavily changing the way the survey works and
the layout of various pages. This commit is the first step, but there's
still a lot of mess to clean up.