simplify travel plan validations and factory
The 'bad_flight_count' validation is an unecessary layer of
complication. It should never fail in the real world because all the
validations are being handled at a different level by the NewTravelPlan
form object, and it adds complication to the factory and thus makes
testing a pain in the ass.

Remove this validation, and simplify the travel plan factory in a way
that wasn't possible before.