add declined_at to card_accounts
Previously we were storing the 'declined_at' timestamp in the
'applied_at' column and aliasing it, because a CardAccount can never be
both 'declined' and 'applied', so this would save database space.

However, the overhead of an extra DB column is tiny and I don't think
it's worth adding an extra layer of complication - however small - for
the developer to keep in his head.

Add declined_at as a separate column