extract HiddenFieldTag component
Note that to make this work I had to change how the 'action' variable is
set in the CardAccount*Action components. Previously 'action' was
undefined in some cases, but this would now raise a warning because
HiddenFieldTag requires the 'action' property to be present. So I
changed it so that 'action' will now be an empty string instead of
undefined in these cases.