change logic of readiness_given_at
The previous approach doesn't work when someone initially gives their status
as not ready, then updates it later - the dashboard will still say 'you said
you were ready on (date you said you were not ready)'. Using updated_at
instead of created_at fixes it for now, but this is still a very
fragile solution. A better idea further down the line would be
to add a column called 'given_at' to the table, or 'not_ready_at'
and 'ready_at', or something like that. But let's defer that to
the future