Updates to card recommendation notes.
1. move the new controller code into `CompleteCardRecommendations`, as
that's where the other logic lives and it all belongs in one place.

I took this opportunity to do away with the concept of a 'service
object', which I've been meaning to remove for a while.
CompleteCardRecommendations is now a Form Object and my edits brings it
in line with the other form objects (or at least most of them).

2. Handle the edge cases covered by the failing tests which were added
in ee054e9f8f.

3. Remove the 'SendNotificationEmail' job as it wasn't actually doing
anything.