remove Eligibility model and table
It was a poor design decision and made everything much more complicated
than necessary.

Instead just use a nullable boolean attribute on 'Person' called
'eligible'.

I want to remove the ReadinessStatus model as well and use a boolean
column on SpendingInfo instead, but that's a more complicated operation
because there are more places within the codebase that interact with
ReadinessStatus (plus 'ready/not_ready_at' is a datum we want to save,
whereas we won't care when someone says they're eligible)