add CardAccount.expirable; refactor CA::Expiration
Remove the dependency of the expiration logic on the scope 'visible'.
From a business logic perspective, card expiration doesn't  necessarily
care whether the card is visible. It's just incidental that right now
all expirable cards are also not visible; this may change further down
the line. So make the 'expirable' scope more explicit to be more
future-proof