reduce OnboardingFlow's responsibilities
This class was doing too much. We can leave the actual persistence to
form objects and/or the controller layer, plus the previous approach
ended up with code like this:

    account.onboarding_survey.add_home_airports!

... which is problematic because a) it's a law of Demeter violation and
b) to an untrained observer, this line appears to be doing something to
the `onboarding_survey`, but actually the method `add_home_airports!` is
doing something to the `account` (namely updating its attributes), which
is very unintuitive.

Similarly, the Account model shouldn't be so tightly coupled to
OnboardingFlow. (And controllers etc should just call
`account.onboarded?`, never `account.onboarding_survey.complete?`)