change sidebar-displaying logic
There's no need to 'manually' hide the sidebar on a view-by-view basis -
the logic is actually simple: show the sidebar if an admin or an
onboarded user is logged in.

This also fixes a bug where some navbar elements were hidden when
there's no sidebar. The only thing we actually need to hide from the
navbar when there's no sidebar is the 'horizontal lines' thing that
shows the sidebar.