big updates to travel plan form
change how 'from' and 'to' save. The 'populate hidden field with JS'
approach was too fiddly, and had bugs: if you moved onto the next field
by pressing tab then sometimes the hidden field wouldn't get updated.
Also it was a PITA making the form fields repopulate correctly after an
unsuccessful submission.

In the new approach we just use the text field directly, and we parse
the airport code from it.