convert AccountTypeForm from react to ERB/jQuery
This is a super-simple page and it's complete pointless building it in
Raect - as shown by the fact that this new setup accomplishes in 16
lines of jQuery what the old code accomplised in 200+ lines of React.

(Note that this page wasn't always so simple; it used to have much more
functionality, which has since been extracted to other pages. The fact
that it was still in React is a hangover from those days.)