Use Time.zone.now instead of Time.zone
I've just arrived in Colombia and suddenly a bunch of tests are failing
because of the timezone my laptop is in (UTC-5, I believe.) Turns out
that Time.now and Date.today don't play well with time zones.

This is just a quick fix to get things roughly working and make the test
suite green again, but we need to have a closer look at how we're
dealing with time zones and make sure there everything is being saved
and displayed correctly.