remove references to intercom events
We're not tracking intercom events anymore, but we still had a bunch of
commented-out or unused code regarding intercom event tracking. We may
add the event tracking back at some point in the distant future, but it
won't be anytime soon, so it's not worth keeping this code around for
now.

Note that we're still tracking intercom *users*, just not *events*.