update CHANGELOG and add comments to Card::Product
We've made the Card -> Card::Product rename as this is clearer, and it
matches the terminology which the banks use themselves. This is a
precursor to renaming 'CardAccount'. I've never been happy with
'CardAccount' as a name as it causes confusion because we already have
an unrelated model called Account. Also, CardAccount does too much, and
we're probably going to extract out some extra models called
Card::Recommendation and Card::Application. After some discussion, the
best naming setup seems to be to rename Card -> Card::Product as we're
already doing, and rename CardAccount to simply 'Card'.