stop saving survey card dates as end of month
The logic behind this was that, since we don't know the exact date these
card accounts were opened/closed (user only tells us month and year, not
day), it's better to err on the side of the date being too late (end of
month) rather than too early (beginning of the month), because we don't
want to remind people too early to apply for the next card when it's
been the minimum length of time since the last card was closed.

The problem is that this requires too much mental overhead and hackery
when saving/updating cards. Let's leave it out for now, and rethink how
we're going to handle this once we actually add the logic which makes
the aforementioned renewal reminders.