remove all data-changing code from migrations
The migrations contained this code because the nature of the schema
changes meant that live data in the production DB also needed to be
changed when the migration was run. Since these migrations have all been
run already in production, and there's no need to make these data
changes when running the migrations from scratch, all this code can be
removed.

This makes the migrations much less fragile when being run from scratch
by a new dev, since using ActiveRecord models within a migration is
asking for trouble and is discouraged.