remove 'lib' frmo autoload path
It's not a good idea to put this on the autoload path when there are so
many files in it that we would never want to autoload. It's also handy
to have a dir which is on our require path but NOT autoloaded, and
that's what lib should be used for.