extract ValidationErrorsAlert::ActiveModel cell
This should be removed eventually because we should use dry-v for all
contracts.

Note that this is currently only being used in one place; most error
alerts are still using the model_validation_errors partial, but that
should be removed and replaced with this cell in all instances. Also
note that the cell isn't an exact replacement for the partial because it
doesn't have the 'full' option and doesn't let you specify a custom
model name.