rename 'survey cards' spec to 'other cards'
The terminology 'survey cards' exists because originally the only two
ways to add a card to your account were to add it in the onboarding
survey, or be recommended the card by an admin. It's now possible
to add card 'manually' after the onboaridng survey, and sooner rather
than later we want to split 'Card' and 'Recommendation' into two
separate models, so this terminology no longer makes sense and we should
start moving away from it.