refactor IntercomJobs::CreateUser; remove VCR
VCR depends on webmock, and webmock is failing all over the place on
Ruby 2.4.0 (which I'm trying to upgrade to on a different branch).
Rather than figure out what the problem is with Webmock (it probably
just needs to be updated to a newer version), I'd rather just remove it
entirely, along with VCR. They're only being used in one test, which
isn't nearly enough to justify the overhead of keeping them around.

That 'one test' is the test for IntercomJobs::CreateUser, and removing
VCR means refactoring the test, so I took the opportunity to refactor
(i.e. simplify) the entire job so it doesn't depend on the DB at all.