AwardWallet: users can link their accounts
add a new model 'AwardWalletUser' which belongs_to Account.

Users will click a link on balances/index which takes them to the
'connect' page, which is just a static page explaining why they should
connect their Abroaders accounts with AwardWallet. This page has a link
to a page on AwardWallet.com where the user can grant us permission to
read their data on AwardWallet.

After granting permission, they're redirected back to our app with
'userId' as a GET param.  We create an 'unloaded' AwardWalletUser for
the current account with this userId, and queue a background job to use
this userId to pull the rest of the user's data from AwardWallet. The
callback page polls and waits for the BG job to be completed. Once it's
completed, and the full AwardWalletUser data has been loaded, they're
redirected to the settings page where they can assign 'owners' on AW
to people on their Abroaders account.

Once the user is connected to AW, they can't see their Abroaders
balances anymore, and instead the balances page is replaced with a link
to AwardWallet. This is a temporary measure until I've updated the
'real' balances page so that it shows AwardWalletAccounts alongside
regular balances.