cards#index - don't show open cards in 'recs' section
Once a recommended card has been successfully applied for and the
account has been opened, it must disappear from the 'recs' section,
since it's not a rec anymore - it's a card account!

Take this opportunity to remove the badly-named scoped
'not_irreversibly_denied'. It's only being used in the 'unresolved'
scope, so just roll it into there. Similarly, the 'visible' scope
shouldn't exist, it's just a synonym for 'unresolved'.