rename 'apply' to 'find my card'
The button text is already changed, but update the underlying stuff like
the HTML IDs as well. We shouldn't think of this as an 'apply' button
because it doesn't update the 'applied_on' timestamp in the DB

When the user comes back and tells us that they applied, THAT's when we
should call it an 'apply' button (or link or form or whatever).