remove unecessary params from CA::Create spec
product_id doesn't do anything when you pass it in as a param nested
under the 'card_account' key. (I think it did do something originally
but I've removed it.) So stop passing in the pointless param, and get
rid of the fancy `params.merge` stuff because it just makes the specs
harder to follow.