fix failing test for #create_admin macro
we can't test the precise email address because the number in the email
address will depend on how many times create_admin is called in
*previous* tests, not just the current example. So previously this test
was passing when run in isolation but would fail if run after another
test had already called create_admin at least once.

I thought about changing the macro so that it resets the sequence back
to 0 after each example, but on closer inspection that's not how
FactoryGirl's sequence method works so let's keep it as it is for the
sake of consistency.