update specs for card rec filters
There was a bug where card producs with non-airline currencies would
ALWAYS get filtered, and would never re-appear even when you removed all
filters, but it actually got fixed as a side effect of the previous
commit.

This commit refactors the specs for all the filters, and also adds some
extra logic that would have caught the old bug.