stash - see commit message
as of 3rd May (writing this on the plane) I've got non-feature specs
passing, but haven't tried feature specs yet.

Before merging this branch I need to think about every place that data
for a card, offer, and card_product might be created, updated, or
displayed, and how the 'unknown offer' concept affects this.