load font awesome from CDN instead of using gem
I'm trying to get things set up so that we can serve assets from Amazon
Cloudfront, but for some reason it's working with everything except the
fonts. It's probably due to the fact that fonts are loaded from a `url`
call in the .css file; the asset pipeline must handle this differently
in a way that e.g. the rack-cors gem doesn't handle the same way.

There's no strong reason to use the gem over a CDN so let's switch to
the CDN and see if we can make things work with Cloudfront that way.